feature: replace security module
