+++ scaleph-ui-react/src/locales/zh-CN/menu.ts
@@ -54,11 +54,12 @@ export default {
   'menu.stdata.refdataMap': '数据映射',
 
   'menu.admin': '系统管理',
-  'menu.admin.dept': '部门管理',
-  'menu.admin.role': '角色管理',
-  'menu.admin.user': '用户管理',
-  'menu.admin.privilege': '权限管理',
-  'menu.admin.resource.web': 'Web 资源',
+  'menu.admin.security': '安全管理',
+  'menu.admin.security.dept': '部门管理',
+  'menu.admin.security.role': '角色管理',
+  'menu.admin.security.user': '用户管理',
+  'menu.admin.security.privilege': '权限管理',
+  'menu.admin.security.resource.web': 'Web 资源',
   'menu.admin.quartz': '系统任务',
   'menu.admin.dict': '数据字典',
   'menu.admin.setting': '系统设置',
