+++ scaleph-ui-react/src/pages/Admin/Security/Privilege/index.tsx
@@ -2,9 +2,9 @@ import React, {useEffect, useState} from 'react';
 import {Button, Card, Col, Empty, Input, List, Row, Space, Tabs, Tree, TreeProps, Typography,} from 'antd';
 import {useAccess, useIntl} from '@umijs/max';
 import {PRIVILEGE_CODE} from '@/constants/privilegeCode';
-import {DeptService} from '@/services/admin/dept.service';
+import {DeptService} from '@/services/admin/security/dept.service';
 import {PrivilegeService} from '@/services/admin/privilege.service';
-import {RoleService} from '@/services/admin/role.service';
+import {RoleService} from '@/services/admin/security/role.service';
 import {SecDeptTreeNode, SecPrivilegeTreeNode, SecRole} from '@/services/admin/typings';
 import styles from './index.less';
 import {TreeNode} from '@/typings';
