+++ scaleph-ui-react/src/pages/Admin/Security/Resource/Web/components/WebResourceForm.tsx
