+++ scaleph-ui-react/src/pages/Admin/Security/Role/components/WebAssugnRoles.tsx
@@ -1,10 +1,10 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { Button, Card, Form, message, Modal, Space } from 'antd';
-import { useIntl } from '@umijs/max';
 import mainHeight from '@/models/useMainSize';
-import TableTransfer from '@/pages/Admin/Resource/Web/components/TransferTable';
+import TableTransfer from '@/pages/Admin/Security/Resource/Web/components/TransferTable';
 import { SecResourceWeb } from '@/services/admin/typings';
 import { AuthService } from '@/services/auth';
+import { Button, Card, Form, message, Modal, Space } from 'antd';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import { useIntl } from '@umijs/max';
 
 // 定义组件 Props 类型
 interface ModalFormParentProps<T> {
@@ -29,17 +29,14 @@ const WebResourceForm: React.FC<ModalFormParentProps<SecResourceWeb>> = ({
   // 角色表格列配置
   const tableColumns = [
     {
-      dataIndex: 'name',
-      title: '角色名称',
+      dataIndex: 'userName',
+      title: '用户名',
       width: 300,
     },
     {
-      dataIndex: 'status.label',
-      title: '角色状态',
+      dataIndex: 'summary',
+      title: '自我介绍',
       width: 300,
-      render: (text: any, record: { status: { label: boolean | React.ReactChild | React.ReactFragment | React.ReactPortal | null | undefined; }; }) => (
-        <span>{record.status.label}</span>
-      ),
     },
   ];
 
@@ -52,22 +49,23 @@ const WebResourceForm: React.FC<ModalFormParentProps<SecResourceWeb>> = ({
   }
 
   // 合并数组
-  function mergeArrays(array1: any, array2: any): any {
-    array1.forEach((obj: { checkOut: number; }, index: any) => {
+  function mergeArrays(array1: Role[], array2: Role[]): Role[] {
+    array1.forEach((obj, index) => {
       obj.checkOut = 0;
     });
-    array2.forEach((obj: { checkOut: number; }, index: any) => {
+    array2.forEach((obj, index) => {
       obj.checkOut = 1;
     });
     return [...array1, ...array2];
   }
+
   // 异步获取数据
   const fetchData = useCallback(async () => {
     try {
-      const res1 = await AuthService.requestUnauthorizedRoles({ userId: data?.id });
-      const res2 = await AuthService.requestUserAuthorizedRoles({ userId: data?.id });
-      if (res1 && res2) {
-        const mergedArray = mergeArrays(res1, res2);
+      const res1 = await AuthService.requestUnauthorizedUsers({ roleId: data?.id });
+      const res2 = await AuthService.requestAuthorizedUsers({ roleId: data?.id });
+      if (res1?.records && res2?.records) {
+        const mergedArray = mergeArrays(res1.records, res2.records);
         setRoleLists(mergedArray);
       }
     } catch (error) {
@@ -85,7 +83,7 @@ const WebResourceForm: React.FC<ModalFormParentProps<SecResourceWeb>> = ({
   const returnTitle = useMemo(() => {
     return (
       <Space direction="vertical">
-        <span>{` ${intl.formatMessage({ id: `menu.${data?.userName}` })}-${intl.formatMessage({ id: 'app.common.operate.new.roles' })}`}</span>
+        <span>{` ${intl.formatMessage({ id: `${data?.name}` })}-${intl.formatMessage({ id: 'app.common.operate.new.rolesUser' })}`}</span>
       </Space>
     );
   }, [data, intl]);
@@ -105,19 +103,19 @@ const WebResourceForm: React.FC<ModalFormParentProps<SecResourceWeb>> = ({
     async (targetKeys, direction, moveKeys) => {
       const roleIds = moveKeys.map((item: string | number) => +item);
       const params = {
-        userId: data?.id,
-        roleIds: roleIds,
+        roleId: data?.id,
+        userIds: roleIds,
       };
       if (direction === 'right') {
         // 批量为角色绑定用户
-        await AuthService.requestUserRoles(params).then((res) => {
+        await AuthService.rolesUser(params).then((res) => {
           if (res?.success) {
             message.success(intl.formatMessage({ id: 'app.common.operate.edit.success' }), 2);
           }
         });
       } else {
         // 批量为角色解除用户绑定
-        await AuthService.requestDeleteUserRoles(params).then((res) => {
+        await AuthService.deleteRolesUser(params).then((res) => {
           message.success(intl.formatMessage({ id: 'app.common.operate.edit.success' }), 2);
         });
       }
@@ -150,7 +148,7 @@ const WebResourceForm: React.FC<ModalFormParentProps<SecResourceWeb>> = ({
         >
           <TableTransfer
             containerHeight={containerInfo.height}
-            titles={[intl.formatMessage({ id: 'app.common.operate.new.notAccreditRoles' }), intl.formatMessage({ id: 'app.common.operate.new.accreditRoles' })]}
+            titles={[intl.formatMessage({ id: 'app.common.operate.new.notAccreditUser' }), intl.formatMessage({ id: 'app.common.operate.new.accreditUser' })]}
             dataSource={roleLists}
             targetKeys={originTargetKeys}
             showSearch={true}
